Refactor code architecture
