+++ src/components/index.ts
@@ -1,6 +1,3 @@
 export * from "./NavBar/NavBar";
 export * from "./Layout/Layout";
 export * from "./Footer/Footer";
-export * from "./Header/Header";
-export * from "./DataTable/DataTable";
-export * from "./ImageViewer/ImageViewer";
