+++ src/features/Dashboard/component/index.ts
@@ -1,0 +1,7 @@
+export * from "./Header/Header";
+export * from "./DataTable/DataTable";
+export * from "./ImageViewer/ImageViewer";
+export * from "./SaleTag/SaleTag";
+export * from "./StyledTableRow/StyledTableRow";
+export * from "./Table/Table";
+export * from "./TableFilters/TableFilters";
