+++ src/features/Dashboard/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from "./useGetData/useGetData";
